snes-sdk/requirements.txt
# SNES SDK Python Dependencies

# Asset conversion tools
Pillow>=10.0

# Testing
pytest>=7.0
